refactor: user 파일구조 정리 및 필요없는 코드 삭제
