+++ src/main/java/com/leets/commitatobe/domain/user/dto/UserRankResponse.java
@@ -1,0 +1,10 @@
+package com.leets.commitatobe.domain.user.dto;
+
+public record UserRankResponse(
+	String githubId,
+	Integer exp,
+	Integer consecutiveCommitDays,
+	String tierName,
+	Integer ranking//랭킹 추가
+) {
+}
