+++ src/main/java/com/leets/commitatobe/domain/user/dto/UserResponse.java
@@ -1,0 +1,10 @@
+package com.leets.commitatobe.domain.user.dto;
+
+public record UserResponse(
+	GitHubAccessTokenResponse gitHubAccessTokenResponse
+) {
+	public record GitHubAccessTokenResponse(
+		String gitHubAccessToken
+	) {
+	}
+}
