+++ src/main/java/com/leets/commitatobe/domain/user/dto/UserSearchResponse.java
@@ -1,0 +1,10 @@
+package com.leets.commitatobe.domain.user.dto;
+
+public record UserSearchResponse(
+	Integer ranking,
+	String githubId,
+	String tierName,
+	Integer exp,
+	Integer consecutiveCommitDays
+) {
+}
