--- src/main/java/com/leets/commitatobe/domain/user/presentation/dto/request/UserRequest.java
@@ -1,5 +1,0 @@
-package com.leets.commitatobe.domain.user.presentation.dto.request;
-
-public record UserRequest() {
-
-}
