--- src/main/java/com/leets/commitatobe/domain/user/presentation/dto/response/UserResponse.java
@@ -1,14 +1,0 @@
-package com.leets.commitatobe.domain.user.presentation.dto.response;
-
-public record UserResponse(
-        RefreshTokenResponse refreshTokenResponse,
-        GitHubAccessTokenResponse gitHubAccessTokenResponse
-) {
-    public record RefreshTokenResponse(
-            String refreshToken
-    ) {}
-
-    public record GitHubAccessTokenResponse(
-            String gitHubAccessToken
-    ) {}
-}
