--- src/main/java/com/leets/commitatobe/domain/user/presentation/dto/response/UserSearchResponse.java
@@ -1,10 +1,0 @@
-package com.leets.commitatobe.domain.user.presentation.dto.response;
-
-public record UserSearchResponse(
-        Integer ranking,
-        String githubId,
-        String tierName,
-        Integer exp,
-        Integer consecutiveCommitDays
-) {
-}
