--- src/main/java/com/leets/commitatobe/domain/user/usecase/UserCommandService.java
@@ -1,4 +1,0 @@
-package com.leets.commitatobe.domain.user.usecase;
-
-public interface UserCommandService {
-}
